Create SPI_CTRL.vhd
This is a controller for the SPI to picoblaze interface. It tells a data register when to grab shift register values using enable (after the required size of data is input using SCK pulses as a counter), and tells the picoblaze when to grab the data register values using an interrupt.